requirements/base.txt
-f http://pypi.svetlyak.ru

#-e .
-e git+git://github.com/andymccurdy/redis-py.git@b82bee8578a6380527b579948e366ec542685f93#egg=redis
-e git+git://github.com/mitsuhiko/flask.git@50dc2403526c5c5c67577767b05eb81e8fab0877#egg=Flask
-e git+git://github.com/svetlyak40wt/flask-script.git#egg=Flask-Script
-e git+git://github.com/kennethreitz/envoy.git@32a536041f93179fe596cb65e06f0017c6ba147d#egg=envoy

Werkzeug==0.8.2
Flask-Mail==0.6.1

# My Fork from: http://github.com/svetlyak40wt/flask-cache/ was already merged to upstream
#Flask-Cache==0.4.0-art1
-e git+git://github.com/svetlyak40wt/flask-cache.git@master#egg=Flask-Cache
#Flask-Assets==0.7
#-e git+git://github.com/miracle2k/flask-assets.git
-e git+git://github.com/miracle2k/webassets.git@ef4d11e24dceb8f7a60d2987b57d3956c4224caf#egg=webassets
-e git+git://github.com/svetlyak40wt/flask-assets.git#egg=Flask-Assets
Flask-WTF==0.6
jsmin==2.0.2

-e git+git://github.com/mitsuhiko/logbook.git@d5aab55d833073056a8b249d6d4c1aae4f6cbfb5#egg=Logbook
-e git+git://github.com/nvie/rq.git@221dd8a0045ceff2e08161f3aa348466c63de47d#egg=rq

requests==0.10.1
pymongo==2.1.1
anyjson==0.3.1
pytz==2012c
# for unique request ids
uuid==1.30
